V0.3.0

+++ BirthdayManager/src/main/java/application/model/PersonManager.java
@@ -3,10 +3,7 @@
  */
 package application.model;
 
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -15,21 +12,24 @@ import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import javafx.beans.property.SimpleBooleanProperty;
+
 /**
  * @author Noah Ruben
  * @see <a href="https://github.com/SirMoM/BirthdayManager">Github</a>
  */
-public class PersonManager{
+public class PersonManager {
 	final static Logger LOG = LogManager.getLogger(PersonManager.class);
 	private static PersonManager personManagerSingelton = null;
+	public SimpleBooleanProperty changesProperty = new SimpleBooleanProperty(false);
 
 	/**
 	 * Static method to create instance of PersonManager class
 	 *
 	 * @return the only instance {@link PersonManager}
 	 */
-	public static PersonManager getInstance(){
-		if(personManagerSingelton == null){
+	public static PersonManager getInstance() {
+		if (personManagerSingelton == null) {
 			personManagerSingelton = new PersonManager();
 		}
 		return personManagerSingelton;
@@ -43,25 +43,27 @@ public class PersonManager{
 	/**
 	 * Private constructor restricted to this class itself.
 	 */
-	private PersonManager(){
+	private PersonManager() {
 		this.setPersonDB(new ArrayList<Person>());
 	}
 
-	public void addNewPerson(final Person newPerson){
+	public void addNewPerson(final Person newPerson) {
 		this.personDB.add(newPerson);
-		if(this.writeThru){
-			this.saveToFile();
-		}
+		changesProperty.set(true);
+//		if (this.writeThru) {TODO DELETE THIS
+//			boolean succsesfull = this.saveToFile();
+//		}
+
 	}
 
 	/**
-	 * Checks if the Save-File exists. If not create the File.
+	 * Checks if the Save-File exists. If not create the File. TODO DELETE THIS
 	 */
-	private void checkSaveFile(){
-		if(!this.saveFile.exists()){
-			try{
+	private void checkSaveFile() {
+		if (!this.saveFile.exists()) {
+			try {
 				this.saveFile.createNewFile();
-			} catch (final IOException ioException){
+			} catch (final IOException ioException) {
 				LOG.catching(ioException);
 			}
 		}
@@ -70,79 +72,64 @@ public class PersonManager{
 	/**
 	 * @param person
 	 */
-	public void deletePerson(final Person person){
+	public void deletePerson(final Person person) {
 		this.personDB.remove(person);
-		if(this.writeThru){
-			this.saveToFile();
-		}
+		this.changesProperty.set(true);
+//		if (this.writeThru) {TODO DELETE THIS
+//			this.saveToFile();
+//		}
 	}
 
-	public Person get(final int indexPerson){
+	public Person getPersonFromIndex(final int indexPerson) {
 		return this.personDB.get(indexPerson);
 	}
 
 	/**
 	 * @return the personDB == the {@link ArrayList} which contains the Persons
 	 */
-	public List<Person> getPersonDB(){
+	public List<Person> getPersons() {
 		return this.personDB;
 	}
 
 	/**
+	 * TODO DELETE THIS
+	 * 
 	 * @return the saveFile
 	 */
-	public File getSaveFile(){
+	public File getSaveFile() {
 		return this.saveFile;
 	}
 
 	/**
-	 * Populates the {@link ArrayList} which contains the Persons ==
-	 * {@link #personDB}
-	 *
-	 * @throws FileNotFoundException if the File is not found
-	 * @throws IOException           if a new File could not be read
-	 */
-	private void populateList() throws IOException, FileNotFoundException{
-		this.personDB.clear();
-		String line;
-		final BufferedReader bufferedReader = new BufferedReader(new FileReader(this.saveFile));
-		while ((line = bufferedReader.readLine()) != null){
-			final Person parsedPerson = Person.parseFromTXTLine(line);
-			if(parsedPerson != null){
-				this.personDB.add(parsedPerson);
-			}
-		}
-		bufferedReader.close();
-	}
-
-	/**
-	 * Saves the {@link ArrayList} {@link #personDB} to the {@link #saveFile}
+	 * Saves the {@link ArrayList} {@link #personDB} to the {@link #saveFile} TODO
+	 * DELETE THIS
 	 *
 	 * @return a boolean which indicates if the save was completed
 	 */
-	public boolean save(){
+	public boolean save() {
 		return this.saveToFile();
 	}
 
 	/**
-	 * Saves the {@link ArrayList} {@link #personDB} to the selectedFile
+	 * Saves the {@link ArrayList} {@link #personDB} to the selectedFile TODO DELETE
+	 * THIS
 	 *
 	 * @param selectedFile a new saveFile
 	 * @return a boolean which indicates if the save was completed
 	 */
-	public boolean save(final File selectedFile){
-		try{
+	public boolean save(final File selectedFile) {
+		try {
 			final FileWriter fileWriter = new FileWriter(selectedFile);
-			if(this.saveFile.getAbsolutePath().endsWith(".csv")){
+			if (this.saveFile.getAbsolutePath().endsWith(".csv")) {
 				LOG.debug("Saving to CSV");
-				for(final Person person : this.personDB){
+				for (final Person person : this.personDB) {
 					fileWriter.write(person.toCSVString());
 					fileWriter.write(System.lineSeparator());
 					fileWriter.flush();
 				}
-			} else{
+			} else {
 				LOG.debug("Saving to TXT");
-				for(final Person person : this.personDB){
+				for (final Person person : this.personDB) {
 					fileWriter.write(person.toTXTString());
 					fileWriter.write(System.lineSeparator());
 					fileWriter.flush();
@@ -150,7 +137,7 @@ public class PersonManager{
 
 			}
 			fileWriter.close();
-		} catch (final IOException ioException){
+		} catch (final IOException ioException) {
 			LOG.catching(ioException);
 			return false;
 		}
@@ -158,28 +145,32 @@ public class PersonManager{
 	}
 
 	/**
-	 * Private saves the {@link ArrayList} {@link #personDB} to the
+	 * Private saves the {@link ArrayList} {@link #personDB} to the TODO DELETE THIS
 	 * {@link #saveFile}
 	 *
 	 * @return a boolean which indicates if the save was completed
 	 */
-	private boolean saveToFile(){
+	private boolean saveToFile() {
+		if (saveFile == null) {
+			return false;
+		}
+
 		this.saveFile.delete();
 		LOG.info("Deleted File", this.saveFile);
 		this.checkSaveFile();
 
-		try{
+		try {
 			final FileWriter fileWriter = new FileWriter(this.saveFile);
-			if(this.saveFile.getAbsolutePath().endsWith(".csv")){
+			if (this.saveFile.getAbsolutePath().endsWith(".csv")) {
 				LOG.debug("Saving to CSV");
-				for(final Person person : this.personDB){
+				for (final Person person : this.personDB) {
 					fileWriter.write(person.toCSVString());
 					fileWriter.write(System.lineSeparator());
 					fileWriter.flush();
 				}
-			} else{
+			} else {
 				LOG.debug("Saving to TXT");
-				for(final Person person : this.personDB){
+				for (final Person person : this.personDB) {
 					fileWriter.write(person.toTXTString());
 					fileWriter.write(System.lineSeparator());
 					fileWriter.flush();
@@ -187,7 +178,7 @@ public class PersonManager{
 
 			}
 			fileWriter.close();
-		} catch (final IOException ioException){
+		} catch (final IOException ioException) {
 			LOG.catching(ioException);
 			return false;
 		}
@@ -197,36 +188,30 @@ public class PersonManager{
 	/**
 	 * @param personDB the personDB to set
 	 */
-	public void setPersonDB(final List<Person> personDB){
+	public void setPersonDB(final List<Person> personDB) {
 		this.personDB = personDB;
 	}
 
 	/**
-	 * @param saveFile the saveFile to set
+	 * @param saveFile the saveFile to set TODO DELETE THIS
 	 */
-	public void setSaveFile(final File saveFile){
+	public void setSaveFile(final File saveFile) {
 		this.saveFile = saveFile;
-		if(saveFile == null){
+		if (saveFile == null) {
 			LOG.info("Warn saveFile is null");
-		} else if(saveFile.exists()){
-			try{
-				this.populateList();
-			} catch (final IOException ioException){
-				LOG.catching(ioException);
-			}
-		} else{
-			try{
+		} else if (!saveFile.exists()) {
+			try {
 				saveFile.createNewFile();
-			} catch (final IOException ioException){
+			} catch (final IOException ioException) {
 				LOG.catching(ioException);
 			}
 		}
 	}
 
 	/**
-	 * @param saveFile the saveFile to set
+	 * @param saveFile the saveFile to set TODO DELETE THIS
 	 */
-	public void setSaveFilePlain(final File saveFile){
+	public void setSaveFilePlain(final File saveFile) {
 		this.saveFile = saveFile;
 	}
 
@@ -234,13 +219,14 @@ public class PersonManager{
 	 * @param indexPerson   the index of the person which will be updated
 	 * @param updatedPerson the person which was updated
 	 */
-	public void updatePerson(final int indexPerson, final Person updatedPerson){
+	public void updatePerson(final int indexPerson, final Person updatedPerson) {
 		this.personDB.get(indexPerson).setBirthday(updatedPerson.getBirthday());
 		this.personDB.get(indexPerson).setName(updatedPerson.getName());
 		this.personDB.get(indexPerson).setSurname(updatedPerson.getSurname());
 		this.personDB.get(indexPerson).setMisc(updatedPerson.getMisc());
-		if(this.writeThru){
-			this.saveToFile();
-		}
+		this.changesProperty.set(true);
+//		if (this.writeThru) {
+//			this.saveToFile();
+//		}
 	}
 }
